pyproject.toml
[project]
name = "automation-server-client"
version = "0.3.0"
description = "This implements the base classes for building processes for the automation server."
readme = "README.md"
authors = [
    { name = "Michael Wulff Nielsen", email = "miwn@odense.dk" }
]
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "pydantic>=2.5.0",
    "httpx>=0.28.1",
]

[project.urls]
homepage = "https://github.com/odense-rpa/automation-server"
repository = "https://github.com/odense-rpa/automation-server"
documentation = "https://github.com/odense-rpa/automation-server/docs"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.12.4",
]

[[tool.hatch.license]]
name = "MIT"
